+++ inst/examples/multiple_mab_simulation_example.R
@@ -5,7 +5,6 @@ tanf <- tanf[1:50, ]
 
 # The seeds passed must be integers, so it is highly recommended to create them
 # before using `sample.int()`
-set.seed(1)
 seeds <- sample.int(10000, 5)
 conditions <- c("no_letter", "open_appt", "specific_appt")
 
@@ -26,19 +25,9 @@ x <- multiple_mab_simulation(
     id_col = "ic_case_id",
     success_col = "success"
   ),
-  verbose = FALSE, times = 5, seeds = seeds, keep_data = TRUE
+  verbose = FALSE, times = 5, seeds = seeds, keep_data = FALSE
 )
 print(x)
-
-# Its Recommenced to set keep_data at FALSE unless necessary to avoid
-# the output from taking up to much memory
-# Keep TRUE
-object.size(x)
-
-x$final_data_nest <- NULL
-# Size if Keep was FALSE
-object.size(x)
-
 
 ## Parallel Execution using future:
 ## Check the future and furrr documentation for more details on possible options